+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # MARSEILLE BB
 
+## WIP v1
+
+- requests
+  - browse through request
+  - requests has random constraints (activity, crowd)
+  - requests get harder each day
+- activities
+  - can schedule each roommates' day
+  - compute the happiness gained/lost and energy consumed at the end of the day
+- days
+  - browse through past days and activities
+- action
+  - fire roommate (big malus but gambling on the happiness lose)
+
 ## v0.1
 
 ![v0.1](docs/v0.1.png)